requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.1
idna==3.4
multidict==6.0.2
nats-py==2.2.0
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.5.1
